Moved files into a PaintBrush directory to make them easier to relocate later.
